In-progress version of mergeLocalizations tool
This performs the same function as the "Run Script to Proof
Localizations" Ruby build script (which never worked that well). This
re-write is in Objective-C, and its main advantage is that it can merge
format specifier changes.

So if we had, for lolcats.lproj:
  "This is version %d." = "I can haz tha %d version.";
  "%@ version %d." = "Itz version %2$d of mah app %1$@.";

And we changed the version in source from 32-bit int to a 64-bit long
(say, because we went 64-bit), the localization _should_ change to:
  "This is version %lu." = "I can haz tha %lu version.";
  "%@ version %lu." = "Itz version %2$lu of mah app %1$@.";

To do that trickery, it also parses format strings, causing me to
already pick up an invalid one.